Split up the utils module
To stop utils.py becoming a dumping ground, split the current code
into json, chunking, and openai sub-modules.
