+++ pyproject.toml
@@ -95,7 +95,7 @@ known-local-folder = ["tuning"]
 disable_error_code = ["import-not-found", "import-untyped"]
 exclude = [
     "^src/instructlab/sdg/generate_data\\.py$",
-    "^src/instructlab/sdg/utils\\.py$",
+    "^src/instructlab/sdg/utils/openai\\.py$",
     "^src/instructlab/sdg/default_flows\\.py$",
     "^src/instructlab/sdg/llmblock\\.py$",
     "^src/instructlab/sdg/utilblocks\\.py$",
