+++ src/instructlab/sdg/utils/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-License-Identifier: Apache-2.0
+
+# This is part of the public API, and used by instructlab
+class GenerateException(Exception):
+    """An exception raised during generate step."""
