+++ tests/test_chunking.py
@@ -1,26 +1,21 @@
 # SPDX-License-Identifier: Apache-2.0
 
-# Standard
-from unittest.mock import Mock, patch
-
 # Third Party
-import git
 import pytest
-import yaml
 
 # First Party
-from instructlab.sdg import utils
+from instructlab.sdg.utils import chunking
 
 # Local
 from .testdata import testdata
 
 
-class TestUtils:
-    """Test collection in instructlab.utils."""
+class TestChunking:
+    """Test collection in instructlab.utils.chunking."""
 
     def test_chunk_docs_wc_exeeds_ctx_window(self):
         with pytest.raises(ValueError) as exc:
-            utils.chunk_document(
+            chunking.chunk_document(
                 documents=testdata.documents,
                 chunk_word_count=1000,
                 server_ctx_size=1034,
@@ -32,7 +27,7 @@ class TestUtils:
 
     def test_chunk_docs_chunk_overlap_error(self):
         with pytest.raises(ValueError) as exc:
-            utils.chunk_document(
+            chunking.chunk_document(
                 documents=testdata.documents,
                 chunk_word_count=5,
                 server_ctx_size=1034,
@@ -44,14 +39,14 @@ class TestUtils:
 
     def test_chunk_docs_long_lines(self):
         chunk_words = 50
-        chunks = utils.chunk_document(
+        chunks = chunking.chunk_document(
             documents=testdata.long_line_documents,
             chunk_word_count=chunk_words,
             server_ctx_size=4096,
         )
-        max_tokens = utils.num_tokens_from_words(chunk_words)
-        max_chars = utils.num_chars_from_tokens(max_tokens)
-        max_chars += utils.DEFAULT_CHUNK_OVERLAP  # add in the chunk overlap
+        max_tokens = chunking._num_tokens_from_words(chunk_words)
+        max_chars = chunking.num_chars_from_tokens(max_tokens)
+        max_chars += chunking._DEFAULT_CHUNK_OVERLAP  # add in the chunk overlap
         max_chars += 50  # and a bit extra for some really long words
         for chunk in chunks:
             assert len(chunk) <= max_chars
